+++ CHANGELOG.md
@@ -21,6 +21,8 @@ The format is based on [Keep a Changelog] and this project adheres to
 - `AnyObject::is_{nil|undefined|true|false}` to `const` functions
 - `Exception::raise` return type to `!`
 - `Mixin::attr` and friends to `Mixin::def_attr`
+- `Array::{contains|remove_all}` to take an arg that implements `Into<O>`
+- `Array::remove_all` to return `Option<O>`
 
 ## [0.0.8] - 2019-05-23
 ### Added